Extract option funcs and response types into separate files for clarity
